32/64-bit printf fixes. fix std::min() usage in ide.cpp that caused compile failures on 64-bit. Also reminder to self: on x86_64, unsigned int is still 32-bit but Bitu is 64-bit, which is why many printf statements in this code have problems on 64-bit.

+++ src/cpu/callback.cpp
@@ -520,7 +520,7 @@ Bitu CALLBACK_SetupExtra(Bitu callback, Bitu type, PhysPt physAddress, bool use_
 		phys_writeb(physAddress+0x04,(Bit8u)0xCF);		//An IRET Instruction
 		return (use_cb?9:5);
 	default:
-		E_Exit("CALLBACK:Setup:Illegal type %d",type);
+		E_Exit("CALLBACK:Setup:Illegal type %u",(unsigned int)type);
 	}
 	return 0;
 }
